Use maybe_queryset in DjangoFilterConnectionField
In our projects we've found a need to inherit from
DjangoFilterConnectionField in order to customize the queryset. When
passing the queryset into "connection_resolver", an error was
raised about "get_queryset" not existing. This is due to the
assumption that the "base_manager" is always a manager. Using
"maybe_queryset" resolves the issue.